IftU: fix some outstanding graphic glitches with the dark gate terrain
We fix this by using the older SHEX-based definition from AtS, although
the mainline macro was marked as deprecated by mog in 1.5.x. If it gets
removed, we'll need to redefine it ourselves.

(I don't remember if this method had its own glitches - I can't see any
problem in the editor at least.)