SoW: rework the help system
The help can now also be viewed cut into partswhen hovering over the unit
type names of the leaders, which are alway present for side 1-9.
This means it can be viewed out-of-turn and without blocking the turn bell.
Remove the image from the right-click menu trigger since pango seems
more important than that image.