AtS: fix fire attack animation glitches affecting Elynia
As with the Goliath case, there are multiple [if]/[else] blocks
interfering with each other. To solve this, one of them has been
eliminated, and the fire impact effect moved to a separate macro used
only when hits=yes, resulting in an accidental code optimization.

It's possible the fire stream effect is glitching in a non-obvious way
since it uses the drake fire breath macro, which has its own [if]/[else]
block within.